requirements.txt
accelerate==0.32.1
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
Brotli==1.0.9
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==2.0.4
click==8.1.7
comm==0.2.2
contourpy==1.2.1
cssselect2==0.7.0
cycler==0.12.1
Cython==3.0.10
datasets==2.20.0
debugpy==1.6.7
decorator==5.1.1
dill==0.3.8
distro==1.9.0
dnspython==2.6.1
email_validator==2.2.0
environs==9.5.0
exceptiongroup==1.2.2
executing==2.0.1
fastapi==0.111.1
fastapi-cli==0.0.4
filelock==3.13.1
fire==0.6.0
FlagEmbedding==1.2.10
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.5.0
gmpy2==2.1.2
grpcio==1.63.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.24.0
idna==3.7
imageio==2.34.2
imgaug==0.4.0
importlib_metadata==8.0.0
importlib_resources==6.4.0
ipykernel==6.29.5
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
lazy_loader==0.4
llvmlite==0.43.0
lmdb==1.5.1
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.21.3
matplotlib==3.9.1
matplotlib-inline==0.1.7
mdurl==0.1.2
milvus-lite==2.4.8
minio==7.2.7
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
nest_asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
numba==0.60.0
numpy==1.26.4
openai==1.36.1
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt-einsum==3.3.0
packaging==24.1
paddleocr==2.8.1
paddlepaddle==2.6.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pip==24.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
protobuf==5.27.2
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pyarrow-hotfix==0.6
pyclipper==1.3.0.post5
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
pydyf==0.11.0
Pygments==2.18.0
pymilvus==2.4.4
pynndescent==0.5.13
pyparsing==3.1.2
pypdfium2==4.30.0
pyphen==0.15.0
PySocks==1.7.1
python-dateutil==2.9.0
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
rapidfuzz==3.9.4
regex==2024.5.15
requests==2.32.2
rich==13.7.1
safetensors==0.4.3
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.13.1
sentence-transformers==3.0.1
setuptools==69.5.1
shapely==2.0.5
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.2
starlette==0.37.2
sympy==1.12
termcolor==2.4.0
threadpoolctl==3.5.0
tifffile==2024.7.21
tinycss2==1.3.0
tokenizers==0.19.1
torch==2.3.0
torchaudio==2.3.0
torchvision==0.18.0
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.42.4
triton==2.3.0
typer==0.12.3
typing_extensions==4.11.0
tzdata==2024.1
ujson==5.10.0
umap-learn==0.5.6
urllib3==2.2.2
uvicorn==0.30.3
uvloop==0.19.0
watchfiles==0.22.0
wcwidth==0.2.13
weasyprint==62.3
webencodings==0.5.1
websockets==12.0
wheel==0.43.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.19.2
zopfli==0.2.3
